% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agefunctions.R
\name{currentage}
\alias{currentage}
\title{A Current Age Function}
\usage{
currentage(birthyear = currentyear)
}
\arguments{
\item{birthyear}{A four digit number indicating a year of birth?}
}
\value{
Current age in years
}
\description{
This function calculates a current age in years when given a birth year.
}
\examples{
# Calculate the age of someone born in 1980
currentage(1980)

}
\keyword{age}
